\name{readMatrix}
\alias{readMatrix}
\title{Reads Matrix}
\description{This is a data frame with four columns and the column names are "AN", "BN", "AT"
and "BT". They are A-allele coverage in the tumor (case) sample, B-allele coverage
in the tumor (case) sample, A-allele coverage in the normal (control) sample,
and B-allele coverage in the normal (control) sample, respectively.}

